mqtt1MinTick: disable check for watchdog reboot
The current code assumes that an external 'ping' message will
periodically be sent to the trelliswifi device. If 20+ minutes
go by and none is received, the device reboots.

While that is a nice watchdog to ensure MQTT connectivity is
healthy, it may become a nuisance for simple deployments. So
I'm disabling it with this change.